move: article, user 패키지 분리
