+++ src/test/java/kr/codesqaud/cafe/service/UserServiceTest.java
@@ -1,7 +1,7 @@
 package kr.codesqaud.cafe.service;
 
 import kr.codesqaud.cafe.domain.User;
-import kr.codesqaud.cafe.repository.MemoryUserRepository;
+import kr.codesqaud.cafe.repository.user.MemoryUserRepository;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
